Add files via upload
I uploaded engraver.sql, which is a phpMyAdmin SQL Dump of the database. (I execute it after creating a database called engraver and accessing it.) I also uploaded CSCI 440 - Project Report 4.pdf, which is a list of the queries that I want to make a front end/PHP form for.